+++ packages/form-core/src/EventClient.ts
@@ -9,6 +9,11 @@ type ExtractEventNames<T> = T extends `${string}:${infer EventName}`
 export type BroadcastFormState = {
   id: string
   state: AnyFormState
+}
+
+export type BroadcastFormApi = {
+  id: string
+  state: AnyFormState
   options: AnyFormOptions
 }
 
@@ -33,29 +38,20 @@ export type BroadcastFormSubmissionState =
       successful: true
     }
 
-export type BroadcastFormUnmounted = {
-  id: string
-}
-
-export type RequestFormState = {
-  id: string
-}
-
-export type RequestFormReset = {
-  id: string
-}
-
-export type RequestFormForceReset = {
+export type BroadcastFormId = {
   id: string
 }
 
 type EventMap = {
-  'form-devtools:form-state-change': BroadcastFormState
-  'form-devtools:form-submission-state-change': BroadcastFormSubmissionState
-  'form-devtools:form-unmounted': BroadcastFormUnmounted
-  'form-devtools:request-form-state': RequestFormState
-  'form-devtools:request-form-reset': RequestFormReset
-  'form-devtools:request-form-force-submit': RequestFormForceReset
+  'form-devtools:form-state': BroadcastFormState
+  'form-devtools:form-api': BroadcastFormApi
+  'form-devtools:form-submission': BroadcastFormSubmissionState
+
+  'form-devtools:request-form-state': BroadcastFormId
+  'form-devtools:request-form-reset': BroadcastFormId
+  'form-devtools:request-form-force-submit': BroadcastFormId
+
+  'form-devtools:form-unmounted': BroadcastFormId
 }
 
 export type EventClientEventMap = keyof EventMap
@@ -66,6 +62,7 @@ class FormEventClient extends EventClient<EventMap> {
   constructor() {
     super({
       pluginId: 'form-devtools',
+      reconnectEveryMs: 1000,
     })
   }
 }