+++ packages/react-form/src/useField.tsx
@@ -1,4 +1,6 @@
-import React, { useMemo, useRef, useState } from 'react'
+'use client'
+
+import { useMemo, useRef } from 'react'
 import { useStore } from '@tanstack/react-store'
 import { FieldApi, functionalUpdate } from '@tanstack/form-core'
 import { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect'