+++ packages/react-form/src/useForm.tsx
@@ -1,6 +1,8 @@
+'use client'
+
 import { FormApi, functionalUpdate } from '@tanstack/form-core'
 import { useStore } from '@tanstack/react-store'
-import { useState } from 'react'
+import { useId, useState } from 'react'
 import { Field } from './useField'
 import { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect'
 import type {
@@ -182,6 +184,8 @@ export function useForm<
     TSubmitMeta
   >,
 ) {
+  const formId = useId()
+
   const [formApi] = useState(() => {
     const api = new FormApi<
       TFormData,
@@ -196,7 +200,7 @@ export function useForm<
       TOnDynamicAsync,
       TOnServer,
       TSubmitMeta
-    >(opts)
+    >({ ...opts, formId: formId })
 
     const extendedApi: ReactFormExtendedApi<
       TFormData,