+++ packages/solid-form-devtools/src/production.ts
@@ -1,0 +1,5 @@
+export { FormDevtoolsPanel } from './FormDevtools'
+
+export type { FormDevtoolsSolidInit } from './FormDevtools'
+
+export { formDevtoolsPlugin } from './plugin'